change key to int, vendor xref

+++ prime-dotnet-webapi/Configuration/HealthAuthorityOrganizationConfiguration.cs
@@ -11,12 +11,12 @@ namespace Prime.Configuration
             get
             {
                 return new[] {
-                    new HealthAuthorityOrganization { Id = HealthAuthorityCode.NorthernHealth,                    Name = "Northern Health",                      CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE },
-                    new HealthAuthorityOrganization { Id = HealthAuthorityCode.InteriorHealth,                    Name = "Interior Health",                      CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE },
-                    new HealthAuthorityOrganization { Id = HealthAuthorityCode.VancouverCoastalHealth,            Name = "Vancouver Coastal Health",             CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE },
-                    new HealthAuthorityOrganization { Id = HealthAuthorityCode.IslandHealth,                      Name = "Island Health",                        CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE },
-                    new HealthAuthorityOrganization { Id = HealthAuthorityCode.FraserHealth,                      Name = "Fraser Health",                        CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE },
-                    new HealthAuthorityOrganization { Id = HealthAuthorityCode.ProvincialHealthServicesAuthority, Name = "Provincial Health Services Authority", CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE }
+                    new HealthAuthorityOrganization { Id = (int)HealthAuthorityCode.NorthernHealth,                    Name = "Northern Health",                      CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE },
+                    new HealthAuthorityOrganization { Id = (int)HealthAuthorityCode.InteriorHealth,                    Name = "Interior Health",                      CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE },
+                    new HealthAuthorityOrganization { Id = (int)HealthAuthorityCode.VancouverCoastalHealth,            Name = "Vancouver Coastal Health",             CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE },
+                    new HealthAuthorityOrganization { Id = (int)HealthAuthorityCode.IslandHealth,                      Name = "Island Health",                        CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE },
+                    new HealthAuthorityOrganization { Id = (int)HealthAuthorityCode.FraserHealth,                      Name = "Fraser Health",                        CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE },
+                    new HealthAuthorityOrganization { Id = (int)HealthAuthorityCode.ProvincialHealthServicesAuthority, Name = "Provincial Health Services Authority", CreatedTimeStamp = SEEDING_DATE, UpdatedTimeStamp = SEEDING_DATE }
                 };
             }
         }
